09.07 13:44 최정준

+++ dasony/src/main/java/com/ds/dasony/alert/model/service/AlertService.java
@@ -8,6 +8,9 @@ import com.ds.dasony.alert.model.vo.Alert;
 
 public interface AlertService {
 	
-	public List<Alert> getAlertList(long userNo);
+
+	public List<Alert> getAlertList(int userNo);
+
+
 
 }
